requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==3.2
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Markdown==3.4.1
phonenumbers==8.13.6
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
